requirements.txt
streamlit
opencv-python-headless
numpy
pandas
plotly
Pillow
mediapipe
tensorflow
imageio
streamlit-webrtc
matplotlib
uuid